+++ src/CommandParser.java
@@ -30,12 +30,12 @@ public class CommandParser {
     }
 
     private AddCommand parseAddCommand(String line) {
-        long priority = new Scanner(line.substring(1)).nextLong();
+        int priority = new Scanner(line.substring(1)).nextInt();
         return new AddCommand(priority);
     }
 
     private RemoveCommand parseRemoveCommand(String line) {
-        long id = new Scanner(line.substring(1)).nextLong();
+        int id = new Scanner(line.substring(1)).nextInt();
         return new RemoveCommand(id);
     }
 
@@ -44,7 +44,7 @@ public class CommandParser {
     }
 
     private QueryCommand parseQueryCommand(String line) {
-        long id = new Scanner(line.substring(1)).nextLong();
+        int id = new Scanner(line.substring(1)).nextInt();
         return new QueryCommand(id);
     }
 }
