+++ src/RemoveCommandResult.java
@@ -2,9 +2,9 @@
  * @author Shujun Ye
  */
 public class RemoveCommandResult extends CommandResult {
-    private long priority;
-    private long position;
-    public RemoveCommandResult(long priority, long position){
+    private int priority;
+    private int position;
+    public RemoveCommandResult(int priority, int position){
         this.priority = priority;
         this.position = position;
     }
